refactor: separate Responses item conversion from transport
