+++ .changeset/quiet-responses-conversion.md
@@ -1,0 +1,5 @@
+---
+'@openai/agents-openai': patch
+---
+
+refactor: separate Responses item conversion from model transport execution without changing behavior.
